pyproject.toml
[tool.poetry]
name = "circulation-import"
version = "0.0.1"
description = "Set of tools facilitating the process of importing book collections into SimplyE's Circulation Manager by using SFTP protocol"
license = "MIT"
authors = ["Viacheslav Bessonov <viacheslav.bessonov@hilbertteam.com>"]
readme = "README.md"
repository = "https://github.com/vbessonov/circulation-import"
homepage = "https://github.com/vbessonov/circulation-import"

[tool.poetry.dependencies]
python = "^3.7"
watchdog = "^0.10.3"
sqlalchemy = "^1.3.19"
sqlalchemy-repr = "^0.0.2"
typing-inspect = "^0.6.0"
"ruamel.yaml" = "^0.16.10"
attrs = "^20.1.0"
click = "^7.1.2"
paramiko = "^2.7.2"

[tool.poetry.dev-dependencies]
nose = "^1.3.7"
pyfakefs = "^4.1.0"
parameterized = "^0.7.4"
bumpversion = "^0.6.0"
mypy = "^0.782"
pylint = "^2.6.0"
nose-timer = "^1.0.0"
flake8 = "^3.8.3"
pylint-exit = "^1.2.0"
coverage = "^5.2.1"
isort = "^5.5.1"
bump2version = "^1.0.0"

[tool.flake8]
max-line-length = 140

[tool.nosetests]
verbosity = 2

[tool.isort]
force_single_line = true
line_length = 140
known_first_party = "common"
default_section = "THIRDPARTY"
forced_separate = "test_common"
skip = "migrations"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
